+++ PyFin/PricingEngines/SabrFormulaImpl.pyx
@@ -69,7 +69,7 @@ cpdef np.ndarray[double, ndim=1] sabrVolatilities(double[:] strikes,
                      double nu,
                      double rho):
     cdef int i
-    cdef int length = len(strikes)
+    cdef size_t length = len(strikes)
     cdef np.ndarray[double, ndim=1] res = np.empty(length, np.float64)
 
     for i in range(length):
